# backend/services/merchant-service/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "支付平台商户管理服务API文档",
        "title": "Merchant Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@payment-platform.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/channels": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "获取所有渠道配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "创建渠道配置",
                "parameters": [
                    {
                        "description": "渠道配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-platform_merchant-service_internal_service.CreateChannelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/channels/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "获取渠道配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "渠道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "更新渠道配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "渠道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "渠道配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-platform_merchant-service_internal_service.UpdateChannelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "删除渠道配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "渠道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "payment-platform_merchant-service_internal_service.CreateChannelInput": {
            "type": "object",
            "required": [
                "channel",
                "config",
                "merchant_id"
            ],
            "properties": {
                "channel": {
                    "type": "string",
                    "enum": [
                        "stripe",
                        "paypal",
                        "crypto",
                        "adyen",
                        "square"
                    ]
                },
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "is_test_mode": {
                    "type": "boolean"
                },
                "merchant_id": {
                    "type": "string"
                }
            }
        },
...
        "payment-platform_merchant-service_internal_service.UpdateChannelInput": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "is_test_mode": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}